fix: persist achievements immediately on unlock
- unlockAchievement() now calls save() to persist to NVS
- Fixed uint32_t truncation to uint64_t in achievements_menu
- Achievements were being set in RAM but never saved to flash

+++ src/ui/achievements_menu.cpp
@@ -159,7 +159,7 @@ void AchievementsMenu::draw(M5Canvas& canvas) {
     canvas.fillScreen(TFT_BLACK);
     
     // Count unlocked
-    uint32_t unlocked = XP::getAchievements();
+    uint64_t unlocked = XP::getAchievements();
     int unlockedCount = 0;
     for (int i = 0; i < TOTAL_ACHIEVEMENTS; i++) {
         if (unlocked & ACHIEVEMENTS[i].flag) unlockedCount++;
@@ -245,7 +245,7 @@ void AchievementsMenu::drawDetail(M5Canvas& canvas) {
 }
 
 void AchievementsMenu::updateBottomOverlay() {
-    uint32_t unlocked = XP::getAchievements();
+    uint64_t unlocked = XP::getAchievements();
     bool hasIt = (unlocked & ACHIEVEMENTS[selectedIndex].flag) != 0;
     
     if (hasIt) {
